+++ Source/Core/GeometryFilters.js
@@ -990,12 +990,9 @@ define([
         return mesh;
     };
 
-    var w1 = new Cartesian2();
-    var w2 = new Cartesian2();
-    var w3 = new Cartesian2();
-    var u1 = new Cartesian3();
-    var u2 = new Cartesian3();
-    var u3 = new Cartesian3();
+    var normalScratch = new Cartesian3();
+    var normalScale = new Cartesian3();
+    var tScratch = new Cartesian3();
     GeometryFilters.computeTangentAndBinormal = function(mesh) {
         if (typeof mesh === 'undefined') {
             throw new DeveloperError('mesh is required.');
@@ -1023,7 +1020,7 @@ define([
         }
         var st = mesh.attributes.st.values;
         if (mesh.attributes.st.componentsPerAttribute !== 2 || st.length % 2 !== 0) {
-            throw new DeveloperError('mesh.attributes.st.values.length must be a multiple of 3');
+            throw new DeveloperError('mesh.attributes.st.values.length must be a multiple of 2');
         }
 
         var indexLists = mesh.indexLists;
@@ -1041,57 +1038,49 @@ define([
 
             var numVertices = mesh.attributes.position.values.length/3;
             var numIndices = indices.length;
-            var tan1 = new Array(numVertices);
+            var tan1 = new Array(numVertices*3);
+
             for (var i = 0; i < tan1.length; i++) {
-                tan1[i] = new Cartesian3();
+                tan1[i] = 0;
             }
 
+            var i03, i13, i23;
             for (i = 0; i < numIndices; i+=3) {
                 var i0 = indices[i];
                 var i1 = indices[i+1];
                 var i2 = indices[i+2];
-                var i03 = i0*3;
-                var i13 = i1*3;
-                var i23 = i2*3;
+                i03 = i0*3;
+                i13 = i1*3;
+                i23 = i2*3;
                 var i02 = i0*2;
                 var i12 = i1*2;
                 var i22 = i2*2;
 
-                u1.x = vertices[i03];
-                u1.y = vertices[i03+1];
-                u1.z = vertices[i03+2];
-                u2.x = vertices[i13];
-                u2.y = vertices[i13+1];
-                u2.z = vertices[i13+2];
-                u3.x = vertices[i23];
-                u3.y = vertices[i23+1];
-                u3.z = vertices[i23+2];
+                var ux = vertices[i03];
+                var uy = vertices[i03+1];
+                var uz = vertices[i03+2];
 
-                w1.x = st[i02];
-                w1.y = st[i02+1];
-                w2.x = st[i12];
-                w2.y = st[i12+1];
-                w3.x = st[i22];
-                w3.y = st[i22+1];
+                var wx = st[i02];
+                var wy = st[i02+1];
+                var t1 = st[i12+1] - wy;
+                var t2 = st[i22+1] - wy;
 
-                var x1 = u2.x - u1.x;
-                var x2 = u3.x - u1.x;
-                var y1 = u2.y - u1.y;
-                var y2 = u3.y - u1.y;
-                var z1 = u2.z - u1.z;
-                var z2 = u3.z - u1.z;
+                var r = 1.0/((st[i12] - wx) * t2 - (st[i22] - wx) * t1);
+                var sdirx = (t2 * (vertices[i13] - ux) - t1 * (vertices[i23] - ux)) * r;
+                var sdiry = (t2 * (vertices[i13+1] - uy) - t1 * (vertices[i23+1] - uy)) * r;
+                var sdirz = (t2 * (vertices[i13+2] - uz) - t1 * (vertices[i23+2] - uz)) * r;
 
-                var s1 = w2.x - w1.x;
-                var s2 = w3.x - w1.x;
-                var t1 = w2.y - w1.y;
-                var t2 = w3.y - w1.y;
+                tan1[i03] += sdirx;
+                tan1[i03+1] += sdiry;
+                tan1[i03+2] += sdirz;
 
-                var r = 1.0/(s1 * t2 - s2 * t1);
-                var sdir = new Cartesian3((t2 * x1 - t1 * x2) * r, (t2 * y1 - t1 * y2) * r, (t2 * z1 - t1 * z2) * r);
+                tan1[i13] += sdirx;
+                tan1[i13+1] += sdiry;
+                tan1[i13+2] += sdirz;
 
-                tan1[i0].add(sdir, tan1[i0]);
-                tan1[i1].add(sdir, tan1[i1]);
-                tan1[i2].add(sdir, tan1[i2]);
+                tan1[i23] += sdirx;
+                tan1[i23+1] += sdiry;
+                tan1[i23+2] += sdirz;
             }
 
             if (typeof mesh.attributes.tangent === 'undefined') {
@@ -1109,22 +1098,28 @@ define([
                 });
             }
             for (i = 0; i < numVertices; i++) {
-                var i3 = i*3;
-                var i31 = i3+1;
-                var i32 = i3+2;
+                i03 = i*3;
+                i13 = i03+1;
+                i23 = i03+2;
 
-                var n = new Cartesian3(normals[i3], normals[i31], normals[i32]);
-                var t = tan1[i].normalize();
+                var n = normalScratch;
+                n.x = normals[i03];
+                n.y = normals[i13];
+                n.z = normals[i23];
+                var t = tScratch;
+                t.x = tan1[i03];
+                t.y = tan1[i13];
+                t.z = tan1[i23];
                 var scalar = n.dot(t);
-                var v = n.multiplyByScalar(scalar);
-                var tr = (t.subtract(v)).normalize();
-                mesh.attributes.tangent.values[i3] = tr.x;
-                mesh.attributes.tangent.values[i31] = tr.y;
-                mesh.attributes.tangent.values[i32] = tr.z;
-                tr = tr.cross(n, tr).normalize();
-                mesh.attributes.binormal.values[i3] = tr.x;
-                mesh.attributes.binormal.values[i31] = tr.y;
-                mesh.attributes.binormal.values[i32] = tr.z;
+                n.multiplyByScalar(scalar, normalScale);
+                t.subtract(normalScale, t).normalize(t);
+                mesh.attributes.tangent.values[i03] = t.x;
+                mesh.attributes.tangent.values[i13] = t.y;
+                mesh.attributes.tangent.values[i23] = t.z;
+                t.cross(n, t).normalize(t);
+                mesh.attributes.binormal.values[i03] = t.x;
+                mesh.attributes.binormal.values[i13] = t.y;
+                mesh.attributes.binormal.values[i23] = t.z;
             }
         }
         return mesh;