+++ CHANGES.md
@@ -17,7 +17,8 @@ Beta Releases
    * Replaced `Uniform.getFrameNumber` and `Uniform.getTime` with `Uniform.getFrameState`, which returns the full frame state.    
 * Improved the performance of drawing polygons created with `configureFromPolygonHierarchy`.
 * Added `GeometryFilters.combine` to combine meshes for better batching.
-* Added `GeometryFilters.computeNormals` to find the normals of vertices in a mesh.
+* Added `GeometryFilters.computeNormal` to find the normals of vertices in a mesh.
+* Added `GeometryFilters.computeTangentAndBinormal` to find the tangent and binormal vectors of vertices in a mesh.
 * Added `BoundingSphere.fromEllipsoid`.
 * Added renderer support for `OES_element_index_uint`, which can improve performance by reducing batch sizes.
 * Added `czm_tangentToEyeSpaceMatrix` built-in GLSL function.